chore: delete `serve` script from baseComponent
There is no need for a `serve` script, since the baseComponent is output as a library. The serve script won’t even work when it is run via `lerna run serve`, because it expects to receive a TTY, and it doesn’t look like lerna passes a TTY to it.