chore: prepare `component-base` for publish
Lerna lets us import our packages as if they were published (e.g. `import <name of export> from @incremental.design/<name of package>`, even if we haven’t actually pushed the packages to npm.

This is important because we want to import our packages in this way when we use them in storybook. It is also literally the point of using Lerna

1. update `packages/vue3/component-base/package.json` with entry points for package
2. update `packages/vue3/component-base/tsconfig.json` to output type declarations
3. add configuration in `packages/vue3/component-base/vue.config.js` to disable `transpileOnly` setting in webpack config and output type declarations