rename `typescript-browser-library` → `typescript-library`
the idea is that vite will bundle libraries just the same regardless of the environment they are meant to run in. However, if you try to run a library that depends on node in the browser, it will just error. Likewise, if you try to run a library that depends on the DOM in node, it will error