remove executors
since we won’t be using them. it’s better to just use shell commands within a container, because then you can manually run the command to see its output. wrapping the commands in an executor just makes it more confusing

refer to the previous commit to see impl of executor, if we need it later on